requirements.txt
customtkinter
mss
Pillow
opencv-python
numpy
clipboard
jaconv
transformers